+++ Sources/Client/Model/Channel/Channel+Requests.swift
@@ -408,4 +408,22 @@ public extension Channel {
     func deleteFile(url: URL, _ completion: @escaping Client.Completion<EmptyData> = { _ in }) -> Cancellable {
         Client.shared.deleteFile(url: url, channel: self, completion)
     }
+    
+    /// Enable slow mode for the channel
+    /// - Parameters:
+    ///   - cooldown: Cooldown duration in seconds. (1-120)
+    ///   - completion: an empty completion block.
+    @discardableResult
+    func enableSlowMode(cooldown: Int,
+                        _ completion: @escaping Client.Completion<EmptyData> = { _ in }) -> Cancellable {
+        Client.shared.enableSlowMode(for: self, cooldown: cooldown, completion)
+    }
+    
+    /// Disables slow mode for the channel
+    /// - Parameters:
+    ///   - completion: an empty completion block.
+    @discardableResult
+    func disableSlowMode(_ completion: @escaping Client.Completion<EmptyData> = { _ in }) -> Cancellable {
+        Client.shared.disableSlowMode(for: self, completion)
+    }
 }